refactor(fintual): move email 2FA policy into the deep module directory
